Remove retrofit and add new execution logic based on result wrappers
Retrofit is unable to provide important information (i.e. the URL)
when converting, as it converts from ResponseBody instead of Response.
Hence it should not be used.
Instead, a home-brewed solution is needed.

AO3Response is a result wrapper that allows the client to handle the
errors in a reasonable and sane way instead of using a hodgepodge of
try-catch blocks, which is hard to reason about.